Move Core.Splash to Windows.SplashWindow
