+++ Apollo/Windows/SplashWindow.xaml
